Update validation rules in Project and Role controllers; add 'id' field to Designer and Project resources; change latitude and longitude fields to decimal in coords migration; enable all CRUD routes for resources

+++ database/migrations/2026_03_04_124533_create_coords_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('coords', function (Blueprint $table) {
             $table->id();
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->decimal('latitude');
+            $table->decimal('longitude');
 
             $table->foreignId('polygon_id')->constrained('polygons')->onDelete('no action');
